* Added the "global" stuff
Reason behind "global" stuff:
  Since we now can include("../dir/any_file.php") at any time, we need to
make sure that all global variables are truly registered as globals.  If I
didn't have them, and within a function, I included a file that was supposed
to set globals, it would not and errors would be generated and improper
execution of the program would result.


git-svn-id: http://squirrelmail.svn.sourceforge.net/svnroot/squirrelmail@1297 7612ce4b-ef26-0410-bec9-ea0150e637f0