Administrator Plugin --------------------
This plugin is not yet complete, but I add it to the CVS in order to get
your coopertaion. Once activated, only the owner of config.php can see
the plugin. Make sure config.php is 660 with the owner set to the one
that can edit it and the group to the use who's running php.


git-svn-id: http://squirrelmail.svn.sourceforge.net/svnroot/squirrelmail@2247 7612ce4b-ef26-0410-bec9-ea0150e637f0