This commit allows the administrator plugin to parse ldap information (not yet to edit it).
Left to go:
* Themes Edit.
* LDAP Edit.
* Data to default (if empty strings or whatever).

Independent mode. For now it is only possible to enter this plugin if
config.php exists, chmod 660, owned by administrator and group = php user.

It should be interesting to hook at login and look for configuration. If
no configuration is found then switch to the plugin to create a deafult one
and start to input data.


git-svn-id: http://squirrelmail.svn.sourceforge.net/svnroot/squirrelmail@2255 7612ce4b-ef26-0410-bec9-ea0150e637f0