Make the base for the SquirrelMail URL configurable. Adds a new variable config_base_location to config.php and a new option to conf.pl. This is to prevent problems in installs where our heuristic doesn't work correctly (#1521299, #1460675, #1110064, #1000850, #1113791)
I believe this is one of the most commonly reported issues in current
stable. That's why I've created the code to be minimally invasive: it only
allows to set the proto+host+port part, because this means all changes
can remain within get_location().

I want to backport this bugfix to stable so I appreciate any feedback
on it. And also of course if you think that this is the wrong way to
solve it :)


git-svn-id: http://squirrelmail.svn.sourceforge.net/svnroot/squirrelmail@11405 7612ce4b-ef26-0410-bec9-ea0150e637f0